+++ src/lib/pages/admin/index.tsx
@@ -25,6 +25,7 @@ import {
   useValidateAddress,
 } from "lib/hooks";
 import { useTxBroadcast } from "lib/providers/tx-broadcast";
+import { AmpEvent, AmpTrack } from "lib/services/amplitude";
 import { queryInstantiateInfo } from "lib/services/contract";
 import type { Addr, ContractAddr, HumanAddr } from "lib/types";
 import { MsgType } from "lib/types";
@@ -89,6 +90,7 @@ const UpdateAdmin = () => {
   });
 
   const proceed = useCallback(async () => {
+    AmpTrack(AmpEvent.ACTION_ADMIN_UPDATE);
     const stream = await updateAdminTx({
       contractAddress: contractAddressParam,
       newAdmin: adminAddress as Addr,