+++ src/lib/pages/contract-details/components/InstantiateInfo.tsx
@@ -121,9 +121,8 @@ export const InstantiateInfo = ({
   }
 
   const instantiatorType = getAddressType(instantiateInfo.instantiator);
-  const adminTypeOpt = instantiateInfo.admin
-    ? getAddressType(instantiateInfo.admin)
-    : undefined;
+  const adminType = getAddressType(instantiateInfo.admin ?? "");
+
   return (
     <Container>
       <LabelText label="Network">{chainId}</LabelText>
@@ -150,13 +149,11 @@ export const InstantiateInfo = ({
 
       <LabelText
         label="Admin Address"
-        helperText1={
-          adminTypeOpt ? getAddressTypeText(adminTypeOpt) : undefined
-        }
+        helperText1={instantiateInfo.admin ? adminType : undefined}
       >
         {instantiateInfo.admin ? (
           <ExplorerLink
-            type={adminTypeOpt}
+            type={adminType}
             value={instantiateInfo.admin}
             canCopyWithHover
           />