+++ src/lib/pages/execute/components/ExecuteArea.tsx
@@ -22,6 +22,7 @@ import { AssetInput, ControllerInput } from "lib/components/forms";
 import JsonInput from "lib/components/json/JsonInput";
 import { useContractStore } from "lib/hooks";
 import { useTxBroadcast } from "lib/providers/tx-broadcast";
+import { AmpEvent, AmpTrack } from "lib/services/amplitude";
 import type { Activity } from "lib/stores/contract";
 import type { ComposedMsg, ContractAddr, HumanAddr, Token } from "lib/types";
 import { MsgType } from "lib/types";
@@ -96,6 +97,7 @@ export const ExecuteArea = ({ control, setValue, cmds }: ExecuteAreaProps) => {
   });
 
   const proceed = useCallback(async () => {
+    AmpTrack(AmpEvent.ACTION_EXECUTE);
     const funds = assets
       .filter((asset) => asset.amount && asset.denom)
       .map((asset) => ({
@@ -179,7 +181,10 @@ export const ExecuteArea = ({ control, setValue, cmds }: ExecuteAreaProps) => {
             <ContractCmdButton
               key={`query-cmd-${cmd}`}
               cmd={cmd}
-              onClickCmd={() => setMsg(jsonPrettify(queryMsg))}
+              onClickCmd={() => {
+                AmpTrack(AmpEvent.USE_CMD_EXECUTE);
+                setMsg(jsonPrettify(queryMsg));
+              }}
             />
           ))}
         </ButtonGroup>