+++ src/lib/pages/instantiate/instantiate.tsx
@@ -24,6 +24,11 @@ import { Stepper } from "lib/components/stepper";
 import WasmPageContainer from "lib/components/WasmPageContainer";
 import { useLCDEndpoint, useValidateAddress } from "lib/hooks";
 import { useTxBroadcast } from "lib/providers/tx-broadcast";
+import {
+  AmpEvent,
+  AmpTrack,
+  AmpTrackToInstantiate,
+} from "lib/services/amplitude";
 import { getCodeIdInfo } from "lib/services/code";
 import type { HumanAddr, Token, U } from "lib/types";
 import { MsgType } from "lib/types";
@@ -155,6 +160,7 @@ const Instantiate = ({ onComplete }: InstantiatePageProps) => {
   // ----------------FUNCTIONS-----------------//
   // ------------------------------------------//
   const proceed = useCallback(() => {
+    AmpTrack(AmpEvent.ACTION_INSTANTIATE);
     handleSubmit(async ({ adminAddress, label, initMsg, assets }) => {
       setSimulating(true);
       const funds = assets
@@ -244,6 +250,10 @@ const Instantiate = ({ onComplete }: InstantiatePageProps) => {
     }
   }, [codeIdQuery, msgQuery, reset, setValue]);
 
+  useEffect(() => {
+    if (router.isReady) AmpTrackToInstantiate(!!msgQuery, !!codeIdQuery);
+  }, [router.isReady, msgQuery, codeIdQuery]);
+
   const validateAdmin = useCallback(
     (input: string) =>
       input && !!validateContractAddress(input) && !!validateUserAddress(input)
@@ -300,7 +310,10 @@ const Instantiate = ({ onComplete }: InstantiatePageProps) => {
                 fontWeight="600"
                 variant="body3"
                 cursor="pointer"
-                onClick={() => setValue("adminAddress", address)}
+                onClick={() => {
+                  AmpTrack(AmpEvent.USE_ASSIGN_ME);
+                  setValue("adminAddress", address);
+                }}
               >
                 Assign me
               </Text>