+++ src/lib/pages/public-project/slug.tsx
@@ -10,11 +10,13 @@ import {
   Icon,
 } from "@chakra-ui/react";
 import { observer } from "mobx-react-lite";
-import { useState } from "react";
+import { useRouter } from "next/router";
+import { useEffect, useState } from "react";
 import { MdExpandMore } from "react-icons/md";
 
 import { CustomTab } from "lib/components/CustomTab";
 import { EmptyState } from "lib/components/state/EmptyState";
+import { AmpEvent, AmpTrack } from "lib/services/amplitude";
 
 import { CodesTable } from "./components/CodesTable";
 import { ContractsTable } from "./components/ContractsTable";
@@ -22,8 +24,14 @@ import { DetailHeader } from "./components/DetailHeader";
 import { usePublicData } from "./data";
 
 export const ProjectDetail = observer(() => {
+  const router = useRouter();
   const [tabIndex, setTabIndex] = useState(0);
   const { publicCodes, publicContracts, projectDetail, slug } = usePublicData();
+
+  useEffect(() => {
+    if (router.isReady) AmpTrack(AmpEvent.TO_PROJECT_DETAIL);
+  }, [router.isReady]);
+
   return (
     <Box py={12} pb={0}>
       <DetailHeader details={projectDetail} slug={slug} />