+++ src/lib/pages/query/components/QueryArea.tsx
@@ -1,4 +1,3 @@
-import * as amplitude from "@amplitude/analytics-browser";
 import { SearchIcon } from "@chakra-ui/icons";
 import { Box, Flex, Spacer, Button, ButtonGroup, Text } from "@chakra-ui/react";
 import { useWallet } from "@cosmos-kit/react";
@@ -13,6 +12,7 @@ import JsonInput from "lib/components/json/JsonInput";
 import JsonReadOnly from "lib/components/json/JsonReadOnly";
 import { DEFAULT_RPC_ERROR } from "lib/data";
 import { useContractStore, useLCDEndpoint, useUserKey } from "lib/hooks";
+import { AmpTrack, AmpEvent } from "lib/services/amplitude";
 import { queryData } from "lib/services/contract";
 import type { ContractAddr, HumanAddr, RpcQueryError } from "lib/types";
 import { encode, getCurrentDate, jsonPrettify, jsonValidate } from "lib/utils";
@@ -67,15 +67,17 @@ export const QueryArea = ({
       },
     }
   );
+  const handleQuery = () => {
+    AmpTrack(AmpEvent.ACTION_QUERY);
+    refetch();
+  };
 
   useEffect(() => setMsg(initialMsg), [initialMsg]);
 
   useEffect(() => {
     const keydownHandler = (e: KeyboardEvent) => {
       // TODO: problem with safari if focusing in the textarea
-      if (e.ctrlKey && e.key === "Enter") {
-        refetch();
-      }
+      if (e.ctrlKey && e.key === "Enter") handleQuery();
     };
     document.addEventListener("keydown", keydownHandler);
     return () => {
@@ -102,7 +104,10 @@ export const QueryArea = ({
               <ContractCmdButton
                 key={`query-cmd-${cmd}`}
                 cmd={cmd}
-                onClickCmd={() => setMsg(jsonPrettify(queryMsg))}
+                onClickCmd={() => {
+                  AmpTrack(AmpEvent.USE_CMD_QUERY);
+                  setMsg(jsonPrettify(queryMsg));
+                }}
               />
             ))}
           </ButtonGroup>
@@ -135,10 +140,7 @@ export const QueryArea = ({
               variant="primary"
               fontSize="14px"
               p="6px 16px"
-              onClick={() => {
-                amplitude.track("Query Clicked");
-                refetch();
-              }}
+              onClick={handleQuery}
               isDisabled={jsonValidate(msg) !== null}
               isLoading={isFetching || isRefetching}
               leftIcon={<SearchIcon />}